Cleanup and move class to appropiate location
